requirements.txt
fastapi[standard]
covalent-api-sdk